requirements.txt
Flask>=1.1.2
numpy>=1.19.1
Flask-SQLAlchemy>=2.4.4